Time: 10 ms (19.01%) | Memory: 49 MB (18.47%) - LeetSync

+++ 1116-maximum-level-sum-of-a-binary-tree/maximum-level-sum-of-a-binary-tree.java
@@ -21,13 +21,13 @@ class Solution {
         Queue<TreeNode>q =new LinkedList<>();
         q.offer(root);
 
-        long maxsum =Integer.MIN_VALUE;
+        int maxsum =Integer.MIN_VALUE;
         int maxlevel =1;
         int currentlevel =1;
 
         while(!q.isEmpty()){
             int levelsize =q.size();
-            long sum =0;
+            int sum =0;
 
             for(int i =0;i<levelsize;i++){
                 TreeNode node =q.poll();
